requirements.txt
flask>=3.0
gunicorn>=22.0
psycopg2-binary>=2.9
feedparser>=6.0
PyJWT>=2.8
requests>=2.31
